Factor scoring metrics out of test.py into a shared scoring_metrics module (used by both test.py and score.py)
score.py was also simplified, and its scope was narrowed.
